move resolving classes
