+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/FlagdProvider.java
@@ -3,6 +3,9 @@ package dev.openfeature.contrib.providers.flagd;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Message;
 import com.google.protobuf.NullValue;
+import dev.openfeature.contrib.providers.flagd.strategy.ResolveStrategy;
+import dev.openfeature.contrib.providers.flagd.strategy.SimpleResolving;
+import dev.openfeature.contrib.providers.flagd.strategy.TracedResolving;
 import dev.openfeature.flagd.grpc.Schema.EventStreamRequest;
 import dev.openfeature.flagd.grpc.Schema.EventStreamResponse;
 import dev.openfeature.flagd.grpc.Schema.ResolveBooleanRequest;
