+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/strategy/SimpleResolving.java
@@ -1,4 +1,4 @@
-package dev.openfeature.contrib.providers.flagd;
+package dev.openfeature.contrib.providers.flagd.strategy;
 
 import com.google.protobuf.Message;
 
@@ -7,7 +7,7 @@ import java.util.function.Function;
 /**
  * {@link SimpleResolving} is a simple request to response resolver.
  */
-class SimpleResolving implements ResolveStrategy {
+public class SimpleResolving implements ResolveStrategy {
 
     @Override
     public <ReqT extends Message, ResT extends Message> ResT resolve(final Function<ReqT, ResT> resolverRef,
