+++ providers/flagd/src/test/java/dev/openfeature/contrib/providers/flagd/TracedResolvingTest.java
@@ -2,6 +2,7 @@ package dev.openfeature.contrib.providers.flagd;
 
 
 import com.google.protobuf.Message;
+import dev.openfeature.contrib.providers.flagd.strategy.TracedResolving;
 import dev.openfeature.flagd.grpc.Schema;
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.trace.Span;
